tog bort gamla checkout och gjorde om gamla payment
